trim unused crate
